+++ administrator/languages/en-GB/com_alfa.ini
@@ -1495,3 +1495,20 @@ COM_ALFA_MEDIA_THUMBNAIL_ERROR = "Please select a valid thumbnail"
 COM_ALFA_MEDIA_SELECT_THUMBNAIL_BUTTON_ERROR_TEXT = "Click to select a thumbnail"
 COM_ALFA_MEDIA_THUMBNAIL_ERROR_ALERT = "One or more media items are missing a valid thumbnail. Please fix the items highlighted in red before saving."
 COM_ALFA_MEDIA_LIMIT_REACHED_ERROR = "Limit reached. File(s) added: "
+
+;Usergroups
+COM_ALFA_HEADING_USERGROUP_ID="Usergroup ID"
+COM_ALFA_HEADING_PRICES_ENABLE="Price Enabled"
+COM_ALFA_USERGROUP_ID_ASC="Usergroup ID ASC"
+COM_ALFA_USERGROUP_ID_DESC="Usergroup ID DESC"
+
+;Users
+COM_ALFA_HEADING_USERNAME="Username"
+COM_ALFA_HEADING_NOTE="Note"
+COM_ALFA_HEADING_LAST_VISIT="Last Visit"
+COM_ALFA_TAB_USER_DETAILS="User Details"
+COM_ALFA_FIELD_NOTE_LABEL="Note"
+COM_ALFA_FIELD_NOTE_DESC="Add a note"
+COM_ALFA_FIELD_JOOMLA_ACCOUNT_LABEL="User's Account"
+COM_ALFA_BTN_EDIT_JOOMLA_USER_POPUP="Edit User's Account"
+COM_ALFA_MODAL_EDIT_USER_TITLE="User's Account Settings"
