+++ administrator/src/Model/UsergroupsModel.php
@@ -1,167 +1,69 @@
 <?php
 
-/**
- * @package    Alfa Commerce
- * @author     Agamemnon Fakas <info@easylogic.gr>
- * @copyright  (C) 2024-2026 Easylogic CO LP / Agamemnon Fakas. All rights reserved.
- * @license    GNU General Public License version 3 or later; see LICENSE
- */
-
 namespace Alfa\Component\Alfa\Administrator\Model;
 
-// No direct access.
 defined('_JEXEC') or die;
 
 use Joomla\CMS\MVC\Factory\MVCFactoryInterface;
 use Joomla\CMS\MVC\Model\ListModel;
 
-/**
- * Methods supporting a list of Usergroups records.
- *
- * @since  1.0.1
- */
 class UsergroupsModel extends ListModel
 {
-    /**
-    * Constructor.
-    *
-     * @param array $config An optional associative array of configuration settings.
-    *
-    * @see        JController
-    * @since      1.6
-    */
     public function __construct($config = [], ?MVCFactoryInterface $factory = null)
     {
         if (empty($config['filter_fields'])) {
             $config['filter_fields'] = [
-                'state', 'a.state',
-                'ordering', 'a.ordering',
-                'created_by', 'a.created_by',
-                'modified_by', 'a.modified_by',
                 'id', 'a.id',
-                'prices_display', 'a.prices_display',
-                'name', 'a.name',
+                'usergroup_id', 'a.usergroup_id',
                 'prices_enable', 'a.prices_enable',
+                'core_title', 'ug.title',
             ];
         }
 
         parent::__construct($config, $factory);
     }
 
-    /**
-     * Method to auto-populate the model state.
-     *
-     * Note. Calling getState in this method will result in recursion.
-     *
-     * @param string $ordering Elements order
-     * @param string $direction Order direction
-     *
-     * @return void
-     *
-     * @throws Exception
-     */
-    protected function populateState($ordering = 'a.id', $direction = 'ASC')
+    protected function populateState($ordering = 'ug.title', $direction = 'ASC')
     {
         parent::populateState($ordering, $direction);
     }
 
-    /**
-     * Method to get a store id based on model configuration state.
-     *
-     * This is necessary because the model is used by the component and
-     * different modules that might need different sets of data or different
-     * ordering requirements.
-     *
-     * @param string $id A prefix for the store id.
-     *
-     * @return string A store id.
-     *
-     * @since   1.0.1
-     */
-    protected function getStoreId($id = '')
-    {
-        // Compile the store id.
-        $id .= ':' . $this->getState('filter.search');
-        $id .= ':' . $this->getState('filter.state');
-
-        return parent::getStoreId($id);
-    }
-
-    /**
-     * Build an SQL query to load the list data.
-     *
-     * @return DatabaseQuery
-     *
-     * @since   1.0.1
-     */
     protected function getListQuery()
     {
-        // Create a new query object.
-        $db = $this->getDbo();
+        $db = $this->getDatabase();
         $query = $db->getQuery(true);
 
-        // Select the required fields from the table.
+        // 1. Select the columns that ACTUALLY exist in your table now
         $query->select(
             $this->getState(
                 'list.select',
-                'DISTINCT a.*',
+                'DISTINCT a.id, a.usergroup_id, a.prices_display, a.prices_enable, ug.title AS core_title',
             ),
         );
-        $query->from('`#__alfa_usergroups` AS a');
+        $query->from($db->quoteName('#__alfa_usergroups', 'a'));
 
-        // Join over the users for the checked out user
-        $query->select('uc.name AS uEditor');
-        $query->join('LEFT', '#__users AS uc ON uc.id=a.checked_out');
+        // 2. JOIN using usergroup_id instead of a.id
+        $query->join('INNER', $db->quoteName('#__usergroups', 'ug') . ' ON ug.id = a.usergroup_id');
 
-        // Join over the user field 'created_by'
-        $query->select('`created_by`.name AS `created_by`');
-        $query->join('LEFT', '#__users AS `created_by` ON `created_by`.id = a.`created_by`');
-
-        // Join over the user field 'modified_by'
-        $query->select('`modified_by`.name AS `modified_by`');
-        $query->join('LEFT', '#__users AS `modified_by` ON `modified_by`.id = a.`modified_by`');
-
-        // Filter by published state
-        $published = $this->getState('filter.state');
-
-        if (is_numeric($published)) {
-            $query->where('a.state = ' . (int) $published);
-        } elseif (empty($published)) {
-            $query->where('(a.state IN (0, 1))');
-        }
-
-        // Filter by search in title
+        // 3. Filter by Search (Cleaned up - removed a.name)
         $search = $this->getState('filter.search');
-
         if (!empty($search)) {
             if (stripos($search, 'id:') === 0) {
                 $query->where('a.id = ' . (int) substr($search, 3));
             } else {
                 $search = $db->Quote('%' . $db->escape($search, true) . '%');
-                $query->where('( a.name LIKE ' . $search . ' )');
+                $query->where($db->quoteName('ug.title') . ' LIKE ' . $search);
             }
         }
 
-        // Add the list ordering clause.
-        $orderCol = $this->state->get('list.ordering', 'id');
+        // 4. Ordering
+        $orderCol = $this->state->get('list.ordering', 'ug.title');
         $orderDirn = $this->state->get('list.direction', 'ASC');
 
-        if ($orderCol && $orderDirn) {
-            $query->order($db->escape($orderCol . ' ' . $orderDirn));
-        }
+        $query->order($db->escape($orderCol . ' ' . $orderDirn));
 
         return $query;
     }
-
-    /**
-     * Get an array of data items
-     *
-     * @return mixed Array of data items on success, false on failure.
-     */
-    public function getItems()
-    {
-        $items = parent::getItems();
-
-        return $items;
-    }
 }
+
+?>
